subscription_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
razorpay==1.3.0
motor==3.3.2
python-dotenv==1.0.0
python-multipart==0.0.6
